add broadcast support to 3 hop shuffle
This commit adds support for broadcasting requests to all processes
using the three hop nexus routing infrastructure.  We add a new API
call:

    hg_return_t shuffle_enqueue_broadcast(shuffle_t sh, uint32_t type, void *d,
                                         uint32_t datalen, int flags);

that enqueues a request for broadcast to all procs except the local proc.
The API supports a flag "SHUFFLE_BCAST_SELF" which causes the local proc
to also get a copy of the data being broadcast (allows the application
to choose the behavior it wants).

To support broadcast, we change the handling of a requests "type" field
by reducing the number of user-defined bits by one.  The new bit
("SHUFFLE_RTYPE_BCAST") is set to tag broadcast requests.  As none of
our apps use that bit, this will not impact current usages of the shuffle.

The following rules apply to broadcast requests:
  - all broadcast requests have "src" set to the global rank of the
    proc that initiates the broadcast

  - all broadcast requests have the "SHUFFLE_RTYPE_BCAST" set in their
    type field.  the delivery callback can use this bit to determine
    if the data being delivered is broadcast or not.

  - all broadcast requests have their "dst" set to the global rank of
    the next process in the broadcast chain.  if the receiving proc
    needs to replicate and forward a request (as part of the broadcast)
    it will update the "dst" to point the new target's global rank.

broadcast requests require handling in 3 places:

[1] when a broadcast request is initiated with the shuffle_enqueue_broadcast()
    call, internally it is translated into a series of normal
    shuffle_enqueue() calls to: ourself (if the "SHUFFLE_BCAST_SELF"
    flag is set), all local procs we talk to (send side of hop 1), and
    all remote procs we talk to (send side of hop 2 when hop 1 is skipped).

[2] when a broadcast request is received from the remote mercury instance
    (recv side of hop 2) we send a copy to ourselves and to any local
    procs we talk to (send side of hop 3).

[3] when a broadcast request is received from the local mercury instance
    there are two possible cases: the receive side of hop 1 and the
    receive side of hop 3.  In both cases we send a copy to ourselves.
    If we are the receive side of hop 1, then in addition we also send
    a copy of the request to all remote nods we talk to (send side of hop 2).

    The code uses nexus and the "src" rank of the request to determine
    if it is the receive side of hop 1 or the receive side of hop 3.
    (in the "hop 1" case, the "src" rank will be on the local node...
     i.e. the "NX_ISLOCAL" case)

In the rpc handler function (shuffle_rpchand()), we now pull requests
from two queues instead of one: the normal list of inbound requests from
the RPC we received, and a new list that contains requests that were
replicated as part of the broadcast process.

The broadcast traffic goes through the normal I/O channels, so all
the current flow control mechanisms apply to it as well.